更新网络

+++ CyxbsMobile2019_iOS/Mine/FeedBackCenter/MainPage/Controllers/FeedBackPage/FeedBackVC.m
@@ -214,6 +214,6 @@
 }
 
 - (void)submit{
-    NSLog(@"%d",self.correctBtn.tag);
+    NSLog(@"%ld",(long)self.correctBtn.tag);
 }
 @end
